Added add and update

+++ Back End/src/main/java/com/example/demo/book/BookController.java
@@ -24,7 +24,7 @@ public class BookController {
     }
 
     @PostMapping
-    public List<Book> addNewBook(@RequestParam("file")MultipartFile file,
+    public List<Book> addNewBook(@RequestParam("invoice")MultipartFile file,
                            @RequestParam("bookName") String bookName,
                            @RequestParam("author") String author,
                            @RequestParam("quantity") int quantity,
@@ -50,20 +50,26 @@ public class BookController {
 
     @PutMapping(path = "{bookId}")
     public Book updateBook(
-            @PathVariable("bookId") Long id,
-            @RequestParam("file")MultipartFile file,
-            @RequestParam("bookName") String bookName,
-            @RequestParam("author") String author,
-            @RequestParam("quantity") int quantity,
-            @RequestParam("price") double price){
+            @PathVariable(required = false, name = "bookId") Long id,
+            @RequestParam(required = false, name = "invoice")MultipartFile file,
+            @RequestParam(required = false, name = "bookName") String bookName,
+            @RequestParam(required = false, name = "author") String author,
+            @RequestParam(required = false, name = "quantity") int quantity,
+            @RequestParam(required = false, name = "price") double price){
 
         String path = null;
+        Book book = bookService.getBook(id);
+        if(book.getInvoicePath() != null && !"".equals(book.getInvoicePath()))
+            path = book.getInvoicePath();
 
-        if(!file.isEmpty()){
-            Book book = bookService.getBook(id);
-            bookService.deleteInvoice(book.getInvoicePath());
+
+        if(file != null){
+            if(book.getInvoicePath() != null && !"".equals(book.getInvoicePath()))
+                bookService.deleteInvoice(book.getInvoicePath());
+
             path = bookService.uploadInvoice(file);
         }
+
         bookService.updateBook(id, bookName,author,price,quantity,path);
 
         return bookService.getBook(id);
@@ -72,7 +78,9 @@ public class BookController {
     @DeleteMapping(path = "{bookId}")
     public Long deleteBook(@PathVariable("bookId") Long id){
         Book book = bookService.getBook(id);
-        bookService.deleteInvoice(book.getInvoicePath());
+        if(book.getInvoicePath() != null && !"".equals(book.getInvoicePath())){
+            bookService.deleteInvoice(book.getInvoicePath());
+        }
         bookService.deleteBook(id);
 
         return id;
